Extract more than just DC images
The list view assumed that the selected image field would contain a list of
Dublin Core formatted JSON (plus an optional special "Specimen" finder for
NHM specimens).
This is altered to allow strings, lists of strings, objects and lists of
objects instead.
If an image delimiter is specified in the resource view settings then this
is taken into account too.