feat: support associatedOccurrences link
This is really an NHM specific change but because of the way the list view is implemented it would be a huge amount of work to implement the hooks and such necessary to modify something at this low level in the processing flow. Given we're going to be moving away from ckanext-list and friends soon and there aren't many users of this, plus this change shouldn't really impact anyone else even if they are using this extension, we'll just implement it here and we can always revert it when we stop using it.